Add more primordials
