+++ llrt_core/src/module_loader/mod.rs
@@ -7,6 +7,7 @@ use once_cell::sync::Lazy;
 use crate::environment;
 
 pub mod loader;
+pub mod require;
 pub mod resolver;
 
 // added when .cjs files are imported
