+++ llrt_core/src/utils/mod.rs
@@ -1,6 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 pub mod clone;
+pub mod ctx;
 pub mod hash;
 pub mod io;
 pub mod latch;
