+++ llrt_core/src/vm.rs
@@ -1,23 +1,8 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-use std::{
-    cmp::min,
-    collections::{HashMap, HashSet},
-    env,
-    ffi::CStr,
-    fmt::Write,
-    fs,
-    process::exit,
-    rc::Rc,
-    result::Result as StdResult,
-    sync::Mutex,
-};
+use std::{cmp::min, env, fmt::Write, process::exit, result::Result as StdResult};
 
 use llrt_json::{parse::json_parse, stringify::json_stringify_replacer_space};
-use llrt_modules::{
-    path::resolve_path,
-    timers::{self, poll_timers},
-};
 use llrt_numbers::number_to_string;
 use llrt_utils::{
     bytes::ObjectBytes,
@@ -31,24 +16,16 @@ use rquickjs::{
     context::EvalOptions,
     function::Opt,
     loader::FileResolver,
-    object::Accessor,
     prelude::{Func, Rest},
-    qjs, AsyncContext, AsyncRuntime, CatchResultExt, CaughtError, Ctx, Error, Filter, Function,
-    IntoJs, Module, Object, Result, Value,
+    AsyncContext, AsyncRuntime, CatchResultExt, CaughtError, Ctx, Error, Function, IntoJs, Object,
+    Result, Value,
 };
-use tokio::time::Instant;
-use tracing::trace;
 
 pub static COMPRESSION_DICT: &[u8] = include_bytes!(concat!(env!("OUT_DIR"), "/compression.dict"));
 
 use crate::{
-    bytecode::BYTECODE_FILE_EXT,
     environment, http,
-    module_loader::{
-        loader::CustomLoader,
-        resolver::{require_resolve, CustomResolver},
-        CJS_IMPORT_PREFIX,
-    },
+    module_loader::{loader::CustomLoader, require, resolver::CustomResolver},
     modules::{console, crypto::SYSTEM_RANDOM, repl::run_repl},
     security,
     utils::clone::structured_clone,
@@ -147,9 +124,8 @@ impl Vm {
                 for init_global in init_globals {
                     init_global(&ctx)?;
                 }
-                timers::init(&ctx)?;
-
-                init(&ctx, module_names)?;
+                init(&ctx)?;
+                require::init(&ctx, module_names)?;
                 Ok(())
             })()
             .catch(&ctx)
@@ -239,7 +215,7 @@ fn json_parse_string<'js>(ctx: Ctx<'js>, bytes: ObjectBytes<'js>) -> Result<Valu
     json_parse(&ctx, bytes)
 }
 
-fn init(ctx: &Ctx<'_>, module_names: HashSet<&'static str>) -> Result<()> {
+fn init(ctx: &Ctx<'_>) -> Result<()> {
     llrt_context::set_spawn_error_handler(|ctx, err| {
         Vm::print_error_and_exit(ctx, err);
     });
@@ -299,204 +275,6 @@ fn init(ctx: &Ctx<'_>, module_names: HashSet<&'static str>) -> Result<()> {
         }),
     )?;
 
-    let require_in_progress: Rc<Mutex<HashMap<Rc<str>, Object>>> =
-        Rc::new(Mutex::new(HashMap::new()));
-
-    // XXX use symbols?
-    const REQUIRE_CACHE_NAME: &str = "__require_cache";
-    const EXPORT_STORE_NAME: &str = "__export_store";
-
-    let require_cache: Object = Object::new(ctx.clone())?;
-    globals.prop(REQUIRE_CACHE_NAME, require_cache)?;
-
-    let exports_store: Object = Object::new(ctx.clone())?;
-    globals.prop(EXPORT_STORE_NAME, exports_store)?;
-
-    let exports_accessor = Accessor::new(
-        |ctx| {
-            let name = get_script_or_module_name(&ctx);
-            let name = name.trim_start_matches(CJS_IMPORT_PREFIX);
-
-            let exports_store: Object = ctx.globals().get(EXPORT_STORE_NAME)?;
-
-            if let Some(value) = exports_store.get_optional::<_, Value>(name)? {
-                Ok::<_, Error>(value.clone())
-            } else {
-                let obj = Object::new(ctx)?.into_value();
-                exports_store.set(name, obj.clone())?;
-                Ok::<_, Error>(obj)
-            }
-        },
-        |ctx, exports| {
-            struct Args<'js>(Ctx<'js>, Value<'js>);
-            let Args(ctx, exports) = Args(ctx, exports);
-            let name = get_script_or_module_name(&ctx);
-            let name = name.trim_start_matches(CJS_IMPORT_PREFIX);
-            let exports_store: Object = ctx.globals().get(EXPORT_STORE_NAME)?;
-            trace!("set module name: {}", name);
-            exports_store.set(name, exports)?;
-            Ok::<_, Error>(())
-        },
-    )
-    .configurable()
-    .enumerable();
-
-    let module = Object::new(ctx.clone())?;
-    module.prop("exports", exports_accessor)?;
-
-    globals.prop("module", module)?;
-    globals.prop("exports", exports_accessor)?;
-
-    globals.set(
-        "require",
-        Func::from(move |ctx, specifier: String| -> Result<Value> {
-            struct Args<'js>(Ctx<'js>);
-            let Args(ctx) = Args(ctx);
-
-            let is_cjs_import = specifier.starts_with(CJS_IMPORT_PREFIX);
-
-            let import_name: Rc<str>;
-
-            let is_json = specifier.ends_with(".json");
-
-            trace!("Before specifier: {}", specifier);
-
-            let import_specifier: Rc<str> = if !is_cjs_import {
-                let is_bytecode = specifier.ends_with(BYTECODE_FILE_EXT);
-                let is_bytecode_or_json = is_json || is_bytecode;
-                let specifier = if is_bytecode_or_json {
-                    specifier
-                } else {
-                    specifier.trim_start_matches("node:").to_string()
-                };
-
-                if module_names.contains(specifier.as_str()) {
-                    import_name = specifier.into();
-                    import_name.clone()
-                } else {
-                    let module_name = get_script_or_module_name(&ctx);
-                    let module_name = module_name.trim_start_matches(CJS_IMPORT_PREFIX);
-                    let abs_path = resolve_path([module_name].iter())?;
-
-                    let resolved_path =
-                        require_resolve(&ctx, &specifier, &abs_path, false)?.into_owned();
-                    import_name = resolved_path.into();
-                    if is_bytecode_or_json {
-                        import_name.clone()
-                    } else {
-                        [CJS_IMPORT_PREFIX, &import_name].concat().into()
-                    }
-                }
-            } else {
-                import_name = specifier[CJS_IMPORT_PREFIX.len()..].into();
-                specifier.into()
-            };
-
-            trace!("After specifier: {}", import_specifier);
-
-            let globals = ctx.globals();
-            let require_cache: Object = globals.get(REQUIRE_CACHE_NAME)?;
-
-            if let Some(cached_value) =
-                require_cache.get::<_, Option<Value>>(import_name.as_ref())?
-            {
-                return Ok(cached_value);
-            }
-
-            if is_json {
-                let json = fs::read_to_string(import_name.as_ref())?;
-                let json = json_parse(&ctx, json)?;
-                require_cache.set(import_name.as_ref(), json.clone())?;
-                return Ok(json);
-            }
-
-            let mut require_in_progress_map = require_in_progress.lock().unwrap();
-            if let Some(obj) = require_in_progress_map.get(&import_name) {
-                let value = obj.clone().into_value();
-                return Ok(value);
-            }
-
-            trace!("Require: {}", import_specifier);
-
-            let obj = Object::new(ctx.clone())?;
-            require_in_progress_map.insert(import_name.clone(), obj.clone());
-            drop(require_in_progress_map);
-
-            let import_promise = Module::import(&ctx, import_specifier.as_bytes())?;
-
-            let rt = unsafe { qjs::JS_GetRuntime(ctx.as_raw().as_ptr()) };
-
-            let mut deadline = Instant::now();
-
-            let mut executing_timers = Vec::new();
-
-            let imported_object = loop {
-                if let Some(x) = import_promise.result::<Object>() {
-                    break x?;
-                }
-
-                if deadline < Instant::now() {
-                    poll_timers(rt, &mut executing_timers, None, Some(&mut deadline))?;
-                }
-
-                ctx.execute_pending_job();
-            };
-
-            let exports_store: Object = globals.get(EXPORT_STORE_NAME)?;
-            let exports_obj: Option<Value> = exports_store.get_optional::<&str, _>(&import_name)?;
-
-            require_in_progress
-                .lock()
-                .unwrap()
-                .remove(import_name.as_ref());
-
-            if let Some(exports_obj) = exports_obj {
-                if exports_obj.type_of() == rquickjs::Type::Object {
-                    let exports = unsafe { exports_obj.as_object().unwrap_unchecked() };
-
-                    for prop in
-                        exports.own_props::<Value, Value>(Filter::new().private().string().symbol())
-                    {
-                        let (key, value) = prop?;
-                        obj.set(key, value)?;
-                    }
-                } else {
-                    //we have explicitly set it
-                    require_cache.set(import_name.as_ref(), exports_obj.clone())?;
-                    return Ok(exports_obj);
-                }
-            }
-
-            let props = imported_object.props::<String, Value>();
-
-            let default_export: Option<Value> = imported_object.get(PredefinedAtom::Default)?;
-            if let Some(default_export) = default_export {
-                //if default export is object attach all named exports to
-                if let Some(default_object) = default_export.as_object() {
-                    for prop in props {
-                        let (key, value) = prop?;
-                        if !default_object.contains_key(&key)? {
-                            default_object.set(key, value)?;
-                        }
-                    }
-                    let default_object = default_object.clone().into_value();
-                    require_cache.set(import_name.as_ref(), default_object.clone())?;
-                    return Ok(default_object);
-                }
-            }
-
-            for prop in props {
-                let (key, value) = prop?;
-                obj.set(key, value)?;
-            }
-
-            let value = obj.into_value();
-
-            require_cache.set(import_name.as_ref(), value.clone())?;
-            Ok(value)
-        }),
-    )?;
-
     //init base primordials
     let _ = BasePrimordials::get(ctx)?;
 
@@ -519,20 +297,4 @@ fn load<'js>(ctx: Ctx<'js>, filename: String, options: Opt<Object<'js>>) -> Resu
     }
 
     ctx.eval_file_with_options(filename, eval_options)
-}
-
-fn get_script_or_module_name(ctx: &Ctx<'_>) -> String {
-    unsafe {
-        let ctx_ptr = ctx.as_raw().as_ptr();
-        let atom = qjs::JS_GetScriptOrModuleName(ctx_ptr, 0);
-        let c_str = qjs::JS_AtomToCString(ctx_ptr, atom);
-        if c_str.is_null() {
-            qjs::JS_FreeCString(ctx_ptr, c_str);
-            return String::from(".");
-        }
-        let bytes = CStr::from_ptr(c_str).to_bytes();
-        let res = std::str::from_utf8_unchecked(bytes).to_string();
-        qjs::JS_FreeCString(ctx_ptr, c_str);
-        res
-    }
 }
